+++ src/services/address/index.ts
@@ -1,9 +1,9 @@
 import * as mock from './mock';
 import * as integration from './integration';
-import { CardPaymentResponse } from '../../lib/interfaces/CardPayment';
+import { AddressResponse } from '../../lib/interfaces/Address';
 
 const useMock = process.env.NODE_ENV === 'mock';
 
-export const getCardPayment: () => Promise<CardPaymentResponse> = useMock
-  ? mock.getCardPayment
-  : integration.getCardPayment;
+export const getAddress: () => Promise<AddressResponse> = useMock
+  ? mock.getAddress
+  : integration.getAddress;
