--- src/services/cardsPayment/integration.ts
@@ -1,6 +1,0 @@
-import { api } from '../api';
-import { AddressResponse } from '../../lib/interfaces/Address';
-
-export async function getAddress(): Promise<AddressResponse> {
-  return api.get('/address');
-}
